Update cleanup of binary snapshot files
cleanup_extra_files gets replaced with cleanup_previous_pending_binary_snapshots
and in cargo-insta we directly remove the relevant binary files now because the
snapshots are already loaded there.
