+++ insta/src/runtime.rs
@@ -2,7 +2,7 @@ use std::cell::RefCell;
 use std::collections::{BTreeMap, BTreeSet};
 use std::error::Error;
 use std::fs;
-use std::io::Write;
+use std::io::{ErrorKind, Write};
 use std::path::{Path, PathBuf};
 use std::rc::Rc;
 use std::str;
@@ -401,6 +401,43 @@ impl<'a> SnapshotAssertionContext<'a> {
         Ok(())
     }
 
+    /// Removes any old .snap.new.* files that belonged to previous pending snapshots. This should
+    /// only ever remove maxium one file because we do this every time before we create a new
+    /// pending snapshot.
+    pub fn cleanup_previous_pending_binary_snapshots(&self) -> Result<(), Box<dyn Error>> {
+        if let Some(ref path) = self.snapshot_file {
+            // The file name to compare against has to be valid utf-8 as it is generated by this crate
+            // out of utf-8 strings.
+            let file_name_prefix = format!("{}.new.", path.file_name().unwrap().to_str().unwrap());
+
+            let read_dir = path.parent().unwrap().read_dir();
+
+            match read_dir {
+                Err(e) if e.kind() == ErrorKind::NotFound => return Ok(()),
+                _ => (),
+            }
+
+            // We have to loop over where whole directory here because there is no filesystem API
+            // for getting files by prefix.
+            for entry in read_dir? {
+                let entry = entry?;
+                let entry_file_name = entry.file_name();
+
+                // We'll just skip over files with non-utf-8 names. The assumption being that those
+                // would not have been generated by this crate.
+                if entry_file_name
+                    .to_str()
+                    .map(|f| f.starts_with(&file_name_prefix))
+                    .unwrap_or(false)
+                {
+                    std::fs::remove_file(entry.path())?;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
     /// Writes the changes of the snapshot back.
     pub fn update_snapshot(
         &self,
@@ -676,9 +713,7 @@ pub fn assert_snapshot(
         assertion_line,
     )?;
 
-    if let Some(snapshot_file) = &ctx.snapshot_file {
-        Snapshot::cleanup_extra_files(ctx.old_snapshot.as_ref(), snapshot_file)?;
-    }
+    ctx.cleanup_previous_pending_binary_snapshots()?;
 
     let content = match new_snapshot_value {
         SnapshotValue::FileText { content, .. } | SnapshotValue::InlineText { content, .. } => {
