+++ insta/src/snapshot.rs
@@ -1,7 +1,7 @@
 use std::env;
 use std::error::Error;
 use std::fs;
-use std::io::{BufRead, BufReader, ErrorKind, Write};
+use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::rc::Rc;
 use std::time::{SystemTime, UNIX_EPOCH};
@@ -597,44 +597,6 @@ impl Snapshot {
         let new_path = path.to_path_buf().with_extension("snap.new");
         self.save_with_metadata(&new_path, &self.metadata)?;
         Ok(new_path)
-    }
-
-    pub fn cleanup_extra_files(snapshot: Option<&Self>, path: &Path) -> Result<(), Box<dyn Error>> {
-        let binary_path = snapshot.and_then(|snapshot| match &snapshot.contents() {
-            SnapshotContents::Text(_) => None,
-            SnapshotContents::Binary(contents) => Some(contents.build_path(path)),
-        });
-        let binary_file_name = binary_path.as_ref().map(|path| path.file_name().unwrap());
-
-        // The file name to compare against has to be valid utf-8 as it is generated by this crate
-        // out of utf-8 strings.
-        let file_name = path.file_name().unwrap().to_str().unwrap();
-
-        let read_dir = path.parent().unwrap().read_dir();
-
-        match read_dir {
-            Err(e) if e.kind() == ErrorKind::NotFound => return Ok(()),
-            _ => (),
-        }
-
-        for entry in read_dir? {
-            let entry = entry?;
-            let entry_file_name = entry.file_name();
-
-            // We'll just skip over files with non-utf-8 names. The assumption being that those
-            // would not have been generated by this crate.
-            if entry_file_name
-                .to_str()
-                .map(|f| f.starts_with(file_name))
-                .unwrap_or(false)
-                && entry_file_name != file_name
-                && binary_file_name.map_or(true, |x| x != entry_file_name)
-            {
-                std::fs::remove_file(entry.path())?;
-            }
-        }
-
-        Ok(())
     }
 }
